Created Button Puzzle
Added code to the player character so that it casts an invisible ray from where the camera is to where it's facing. Used to detect if a button is pressed

Created button objects. Each given an ID number that represents the order they should be pressed. When pressed, they call a function in the Button Manager to determine if the order is correct. Once pressed, a button cannot be pressed again until the manager code decides that the sequence is incorrect.

Created button manager. Whenever a button is pressed, a function is called in the manager to determine if the order is still correct or if the puzzle has been solved.